# openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/lookup/{vin}":{"get":{"summary":"Lookup","operationId":"lookup_lookup__vin__get","parameters":[{"required":true,"schema":{"title":"Vin","type":"string"},"name":"vin","in":"path"},{"required":false,"schema":{"title":"Make","type":"string"},"name":"make","in":"query"},{"required":false,"schema":{"title":"Model","type":"string"},"name":"model","in":"query"},{"required":false,"schema":{"title":"Model Year","type":"string"},"name":"model_year","in":"query"},{"required":false,"schema":{"title":"Body Class","type":"string"},"name":"body_class","in":"query"},{"required":false,"schema":{"title":"Cached","type":"boolean"},"name":"cached","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/remove/{vin}":{"get":{"summary":"Remove","operationId":"remove_remove__vin__get","parameters":[{"required":true,"schema":{"title":"Vin","type":"string"},"name":"vin","in":"path"},{"required":false,"schema":{"title":"Make","type":"string"},"name":"make","in":"query"},{"required":false,"schema":{"title":"Model","type":"string"},"name":"model","in":"query"},{"required":false,"schema":{"title":"Model Year","type":"string"},"name":"model_year","in":"query"},{"required":false,"schema":{"title":"Body Class","type":"string"},"name":"body_class","in":"query"},{"required":false,"schema":{"title":"Cached","type":"boolean"},"name":"cached","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/export":{"get":{"summary":"Export","operationId":"export_export_get","responses":{"200":{"description":"Successful Response"}}}}},"components":{"schemas":{"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}